+++ mse-20190531/alibabacloud_mse20190531/models.py
@@ -2731,7 +2731,9 @@ class AddGatewayAuthRequest(TeaModel):
     def __init__(
         self,
         accept_language: str = None,
+        auth_resource_config: str = None,
         auth_resource_list: List[AddGatewayAuthRequestAuthResourceList] = None,
+        auth_resource_mode: int = None,
         client_id: str = None,
         client_secret: str = None,
         cookie_domain: str = None,
@@ -2753,8 +2755,10 @@ class AddGatewayAuthRequest(TeaModel):
         type: str = None,
     ):
         self.accept_language = accept_language
+        self.auth_resource_config = auth_resource_config
         # The information about the resource to be authorized.
         self.auth_resource_list = auth_resource_list
+        self.auth_resource_mode = auth_resource_mode
         # The application ID registered with the OIDC authentication service.
         self.client_id = client_id
         # The application secret registered with the OIDC authentication service.
@@ -2810,10 +2814,14 @@ class AddGatewayAuthRequest(TeaModel):
         result = dict()
         if self.accept_language is not None:
             result['AcceptLanguage'] = self.accept_language
+        if self.auth_resource_config is not None:
+            result['AuthResourceConfig'] = self.auth_resource_config
         result['AuthResourceList'] = []
         if self.auth_resource_list is not None:
             for k in self.auth_resource_list:
                 result['AuthResourceList'].append(k.to_map() if k else None)
+        if self.auth_resource_mode is not None:
+            result['AuthResourceMode'] = self.auth_resource_mode
         if self.client_id is not None:
             result['ClientId'] = self.client_id
         if self.client_secret is not None:
@@ -2858,11 +2866,15 @@ class AddGatewayAuthRequest(TeaModel):
         m = m or dict()
         if m.get('AcceptLanguage') is not None:
             self.accept_language = m.get('AcceptLanguage')
+        if m.get('AuthResourceConfig') is not None:
+            self.auth_resource_config = m.get('AuthResourceConfig')
         self.auth_resource_list = []
         if m.get('AuthResourceList') is not None:
             for k in m.get('AuthResourceList'):
                 temp_model = AddGatewayAuthRequestAuthResourceList()
                 self.auth_resource_list.append(temp_model.from_map(k))
+        if m.get('AuthResourceMode') is not None:
+            self.auth_resource_mode = m.get('AuthResourceMode')
         if m.get('ClientId') is not None:
             self.client_id = m.get('ClientId')
         if m.get('ClientSecret') is not None:
@@ -2909,7 +2921,9 @@ class AddGatewayAuthShrinkRequest(TeaModel):
     def __init__(
         self,
         accept_language: str = None,
+        auth_resource_config: str = None,
         auth_resource_list_shrink: str = None,
+        auth_resource_mode: int = None,
         client_id: str = None,
         client_secret: str = None,
         cookie_domain: str = None,
@@ -2931,8 +2945,10 @@ class AddGatewayAuthShrinkRequest(TeaModel):
         type: str = None,
     ):
         self.accept_language = accept_language
+        self.auth_resource_config = auth_resource_config
         # The information about the resource to be authorized.
         self.auth_resource_list_shrink = auth_resource_list_shrink
+        self.auth_resource_mode = auth_resource_mode
         # The application ID registered with the OIDC authentication service.
         self.client_id = client_id
         # The application secret registered with the OIDC authentication service.
@@ -2983,8 +2999,12 @@ class AddGatewayAuthShrinkRequest(TeaModel):
         result = dict()
         if self.accept_language is not None:
             result['AcceptLanguage'] = self.accept_language
+        if self.auth_resource_config is not None:
+            result['AuthResourceConfig'] = self.auth_resource_config
         if self.auth_resource_list_shrink is not None:
             result['AuthResourceList'] = self.auth_resource_list_shrink
+        if self.auth_resource_mode is not None:
+            result['AuthResourceMode'] = self.auth_resource_mode
         if self.client_id is not None:
             result['ClientId'] = self.client_id
         if self.client_secret is not None:
@@ -3029,8 +3049,12 @@ class AddGatewayAuthShrinkRequest(TeaModel):
         m = m or dict()
         if m.get('AcceptLanguage') is not None:
             self.accept_language = m.get('AcceptLanguage')
+        if m.get('AuthResourceConfig') is not None:
+            self.auth_resource_config = m.get('AuthResourceConfig')
         if m.get('AuthResourceList') is not None:
             self.auth_resource_list_shrink = m.get('AuthResourceList')
+        if m.get('AuthResourceMode') is not None:
+            self.auth_resource_mode = m.get('AuthResourceMode')
         if m.get('ClientId') is not None:
             self.client_id = m.get('ClientId')
         if m.get('ClientSecret') is not None:
@@ -20333,6 +20357,8 @@ class GetGatewayAuthDetailResponseBodyDataResourceList(TeaModel):
 class GetGatewayAuthDetailResponseBodyData(TeaModel):
     def __init__(
         self,
+        auth_resource_config: str = None,
+        auth_resource_mode: int = None,
         client_id: str = None,
         client_secret: str = None,
         cookie_domain: str = None,
@@ -20358,6 +20384,8 @@ class GetGatewayAuthDetailResponseBodyData(TeaModel):
         token_position: str = None,
         type: str = None,
     ):
+        self.auth_resource_config = auth_resource_config
+        self.auth_resource_mode = auth_resource_mode
         self.client_id = client_id
         self.client_secret = client_secret
         self.cookie_domain = cookie_domain
@@ -20397,6 +20425,10 @@ class GetGatewayAuthDetailResponseBodyData(TeaModel):
             return _map
 
         result = dict()
+        if self.auth_resource_config is not None:
+            result['AuthResourceConfig'] = self.auth_resource_config
+        if self.auth_resource_mode is not None:
+            result['AuthResourceMode'] = self.auth_resource_mode
         if self.client_id is not None:
             result['ClientId'] = self.client_id
         if self.client_secret is not None:
@@ -20451,6 +20483,10 @@ class GetGatewayAuthDetailResponseBodyData(TeaModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('AuthResourceConfig') is not None:
+            self.auth_resource_config = m.get('AuthResourceConfig')
+        if m.get('AuthResourceMode') is not None:
+            self.auth_resource_mode = m.get('AuthResourceMode')
         if m.get('ClientId') is not None:
             self.client_id = m.get('ClientId')
         if m.get('ClientSecret') is not None: